Nuevos cambios

+++ src/mx/itson/catrina/entidades/Movimiento.java
@@ -83,7 +83,7 @@ public class Movimiento {
         return cantidad;
     }
 
-    public void setCantidad(int cantidad) {
+    public void setCantidad(float cantidad) {
         this.cantidad = cantidad;
     }
